+++ lib/presentation/components/app_loading.dart
@@ -1,7 +1,4 @@
 import 'package:count_habits/application/state/loading_state_provider.dart';
-import 'package:count_habits/application/usecase/counter/state/counters_provider.dart';
-import 'package:count_habits/presentation/pages/theme/color_schemes.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -25,33 +22,6 @@ class AppLoading extends ConsumerWidget {
         height: 60,
         width: 60,
         child: CircularProgressIndicator(strokeWidth: 8),
-      ),
-    );
-  }
-}
-
-class ReLoadingWidget extends ConsumerWidget {
-  const ReLoadingWidget({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final theme = ref.watch(cupertinoThemeProvider);
-    return CupertinoPageScaffold(
-      child: Center(
-        child: CupertinoButton(
-          color: theme.barBackgroundColor,
-          onPressed: () => ref.invalidate(countersProvider),
-          borderRadius: BorderRadius.circular(10),
-          child: Text(
-            '再読み込み',
-            style: TextStyle(
-              color: theme.brightness == Brightness.light ? Colors.black : Colors.white,
-              fontSize: 15,
-              fontWeight: FontWeight.w600,
-              fontFamily: theme.textTheme.textStyle.fontFamily,
-            ),
-          ),
-        ),
       ),
     );
   }
